requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
redis
python-dotenv
loguru
httpx
discord.py
passlib[bcrypt]
pika